requirement.txt
attrs==23.1.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
et-xmlfile==1.1.0
exceptiongroup==1.1.3
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-spreadsheet==0.0.6
googleapis-common-protos==1.61.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
pandas==2.1.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.15.2
selenium-stealth==1.0.6
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
wsproto==1.2.0
zstandard==0.22.0
